Move util functions to a seperate folder
