+++ lib/utils/difficultyToColor.tsx
